requirements.txt
accelerate==1.6.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrdict==2.0.1
bitsandbytes==0.45.5
certifi==2025.4.26
charset-normalizer==3.4.2
cmake==4.0.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
-e git+https://github.com/deepseek-ai/DeepSeek-VL.git@681bffb4519856ad27cc17531aacde31ddf6f1a7#egg=deepseek_vl
distro==1.9.0
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
ffmpeg==1.4
filelock==3.18.0
flash_attn @ file:///F:/Proyectos%20Programacion/Proyectos%20Python/VisionSST/req/flash_attn-2.7.4.post1%2Bcu128torch2.7.0cxx11abiFALSE-cp310-cp310-win_amd64.whl#sha256=b83f373b1d8a29d83c7ec53df83061db1b25e2d09e9465d575ffc7444502acf0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipykernel==6.29.5
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.9.0
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.2.5
openai==1.78.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.3
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0+cu126
torchaudio==2.7.0+cu126
torchvision==0.22.0+cu126
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton @ file:///F:/Proyectos%20Programacion/Proyectos%20Python/VisionSST/req/triton-2.0.0-cp310-cp310-win_amd64.whl#sha256=91a6ec395022743269c942df7af01b210f642fb633d146a811be05a455adbae2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
